refactor: decompose EatmeDesktopRunExecutionEvidence from 1103 to 409 lines
Extract evidence writing logic into 5 focused classes:
- BlockerDetail: immutable data record (17 lines)
- PixelObservation: pixel result + JSON factories (93 lines)
- EatmeWindowDetector: component readiness checks (71 lines)
- EatmeScreenshotCapture: Robot capture + PNG write (178 lines)
- EatmeEvidenceWriter: all JSON artifact writing + validation (488 lines)

The coordinator retains: public API, Recorder, constants, and
forwarding delegates preserving all existing test signatures.

All 800 tests pass with zero modifications to existing test files.
